# perf: improve performance of setRows