--- a/packages/toast-ui.grid/types/dataSource/index.d.ts
+++ b/packages/toast-ui.grid/types/dataSource/index.d.ts
@@ -108,7 +108,7 @@ export interface ModifiedDataManager {
   getAllModifiedData: (options: ModifiedRowsOptions) => ModifiedRows;
   isModified: () => boolean;
   isModifiedByType: (type: ModificationTypeCode) => boolean;
-  push: (type: ModificationTypeCode, row: Row, mixed?: boolean) => void;
+  push: (type: ModificationTypeCode, row: Row[], mixed?: boolean) => void;
   clearSpecificRows: (rowMap: MutationParams) => void;
   clear: (type: RequestTypeCode) => void;
   clearAll: () => void;