<!-- EDIT TITLE PLEASE -->
<!-- It should be one of them
  <ISSUE TYPE>: Short Description (<CLOSING TYPE> #<ISSUE NUMBERS>)
  ex)
  feat: add new feature (close #111)
  fix: wrong behavior (fix #111)
  chore: change build tool (ref #111)
-->

<!-- SPECIFY A ISSUE TYPE AT HEAD
  feat: A new feature
  fix: A bug fix
  docs: Documentation only changes
  style: Changes that do not affect the meaning of the code (white-space, formatting etc)
  refactor: A code change that neither fixes a bug or adds a feature
  perf: A code change that improves performance
  test: Adding missing tests
  chore: Changes to the build process or auxiliary tools and libraries such as documentation generation
-->

<!-- ADD CLOSING TYPE AND ISSUE NUMBER AT TAIL
  (<CLOSING TYPE> #<ISSUE NUMBERS>)
  close: resolve not a bug(feature, docs, etc) completely
  fix: resolve a bug completely
  ref: not fully resolved or related to
-->

### Please check if the PR fulfills these requirements
- [x] It's submitted to right branch according to our branching model
- [x] It's right issue type on title
- [ ] When resolving a specific issue, it's referenced in the PR's title (e.g. `fix #xxx[,#xxx]`, where "xxx" is the issue number)
- [x] The commit message follows our guidelines
- [ ] Tests for the changes have been added (for bug fixes/features)
- [ ] Docs have been added/updated (for bug fixes/features)
- [ ] It does not introduce a breaking change or has description for the breaking change

### Description

Improved performance of the setRows API.
1. modified to only make data within the viewport responsive, not all data set by setRows.
2. modified to update summaryData after all data is set, rather than updating summaryData for each row that is set.
3. modifiedDataManager to handle modified data as an array.

This should result in the following performance improvements

Test environment: macOS Ventura 13.1, Chrome 110, 1 million data (10000 rows, 100 columns), using summaryData for all columns

1. using the setRow API: 3,586,784 ms on average (about 1 hour)
2. using setRows API (before performance improvements): 595,712 ms on average (about 10 minutes) 
3. using the setRows API (after performance improvements): 595 ms on average (about 0.6 seconds)

This is a 6028x performance improvement over using the setRow API and a 1001x performance improvement over before the performance improvement.

---
Thank you for your contribution to TOAST UI product. 🎉 😘 ✨